Replace info icons with tile selection
