+++ src/App.js
@@ -10,7 +10,7 @@ import './App.scss';
 import ErrorBoundary from './components/ErrorBoundary';
 import PermissionsChecker from './components/PermissionsChecker';
 import DataLoader from './components/DataLoader';
-import { CLOUD_CARDS_KEY } from './components/CloudCards';
+import { CLOUD_CARDS_KEY } from './components/CloudTiles/CloudCards';
 
 import { getBaseName } from './frontend-components-copies/getBaseName';
