+++ src/test/components/cloudTiles/CloudCards.test.js
@@ -4,15 +4,16 @@ import { act } from 'react-dom/test-utils';
 import { Card, CardHeader } from '@patternfly/react-core';
 import PlusIcon from '@patternfly/react-icons/dist/js/icons/plus-icon';
 
-import CloudCards, { CLOUD_CARDS_KEY, ImageWithPlaceholder } from '../../components/CloudCards';
+import CloudCards, { CLOUD_CARDS_KEY } from '../../../components/CloudTiles/CloudCards';
 
-import componentWrapperIntl from '../../utilities/testsHelpers';
-import { Loader } from '../../components/SourcesTable/loaders';
+import componentWrapperIntl from '../../../utilities/testsHelpers';
+import CloudTiles from '../../../components/CloudTiles/CloudTiles';
 
 describe('CloudCards', () => {
   let wrapper;
   let localStorage;
   let protoTmp;
+  let initialProps;
 
   beforeEach(() => {
     protoTmp = Storage;
@@ -24,6 +25,9 @@ describe('CloudCards', () => {
     Storage.prototype.setItem = jest.fn((name, value) => {
       localStorage[name] = value ? 'true' : 'false';
     });
+    initialProps = {
+      setSelectedType: jest.fn(),
+    };
   });
 
   afterEach(() => {
@@ -32,11 +36,12 @@ describe('CloudCards', () => {
 
   it('renders correctly and sets local storage', async () => {
     await act(async () => {
-      wrapper = mount(componentWrapperIntl(<CloudCards />));
+      wrapper = mount(componentWrapperIntl(<CloudCards {...initialProps} />));
     });
     wrapper.update();
 
     expect(wrapper.find(Card)).toHaveLength(2);
+    expect(wrapper.find(CloudTiles)).toHaveLength(1);
     expect(wrapper.find('img')).toHaveLength(2);
     expect(wrapper.find(PlusIcon)).toHaveLength(3);
 
@@ -52,7 +57,7 @@ describe('CloudCards', () => {
     localStorage[CLOUD_CARDS_KEY] = 'false';
 
     await act(async () => {
-      wrapper = mount(componentWrapperIntl(<CloudCards />));
+      wrapper = mount(componentWrapperIntl(<CloudCards {...initialProps} />));
     });
     wrapper.update();
 
@@ -70,7 +75,7 @@ describe('CloudCards', () => {
 
   it('hides both card at once', async () => {
     await act(async () => {
-      wrapper = mount(componentWrapperIntl(<CloudCards />));
+      wrapper = mount(componentWrapperIntl(<CloudCards {...initialProps} />));
     });
     wrapper.update();
 
@@ -97,26 +102,6 @@ describe('CloudCards', () => {
     expect(wrapper.find(Card).last().props().isExpanded).toEqual(true);
     expect(localStorage).toEqual({
       [CLOUD_CARDS_KEY]: 'true',
-    });
-  });
-
-  describe('ImageWithPlaceholder', () => {
-    it('hides the loader on onLoad', async () => {
-      await act(async () => {
-        wrapper = mount(componentWrapperIntl(<ImageWithPlaceholder src="/some-picture.jpg" />));
-      });
-      wrapper.update();
-
-      expect(wrapper.find(Loader)).toHaveLength(1);
-      expect(wrapper.find('img').props().style.display).toEqual('none');
-
-      await act(async () => {
-        wrapper.find('img').simulate('load');
-      });
-      wrapper.update();
-
-      expect(wrapper.find(Loader)).toHaveLength(0);
-      expect(wrapper.find('img').props().style.display).toEqual('initial');
     });
   });
 });
